+++ src/api/index.js
@@ -1,14 +1,8 @@
 import { Router } from "express"
-import admin from "./routes/admin"
-import store from "./routes/store"
 import hooks from "./routes/hooks"
 
-export default (rootDirectory) => {
+export default (rootDirectory, pluginOptions) => {
   const app = Router()
-
-  store(app, rootDirectory)
-  admin(app, rootDirectory)
-  hooks(app, rootDirectory)
-
+  hooks(app, rootDirectory, pluginOptions)
   return app
 }