+++ src/api/routes/hooks/index.js
@@ -1,18 +1,11 @@
 import { Router } from "express"
 import rateLimit from "express-rate-limit"
 import bodyParser from "body-parser"
-import cors from "cors"
-import { getConfigFile } from "medusa-core-utils"
 import middlewares from "../../middlewares"
 
 const route = Router()
 
-export default (app, rootDirectory) => {
-  const { configModule } = getConfigFile(rootDirectory, "medusa-config")
-  const config = (configModule && configModule.projectConfig) || {}
-
-  const storeCors = config.store_cors || ""
-
+export default (app) => {
   const apiLimiter = rateLimit({
     windowMs: 15 * 60 * 1000,
     max: 100,
@@ -20,20 +13,21 @@ export default (app, rootDirectory) => {
     legacyHeaders: false,
   })
 
-  route.use(
-    cors({
-      origin: storeCors.split(","),
-      credentials: true,
-    })
-  )
-
   app.use("/hooks", route)
   route.use(bodyParser.json())
 
   route.post(
     "/shippo/transaction",
+    middlewares.verifyHook(),
     apiLimiter,
     middlewares.wrap(require("./transactions").default)
+  )
+
+  route.post(
+    "/shippo/track",
+    middlewares.verifyHook(),
+    apiLimiter,
+    middlewares.wrap(require("./track").default)
   )
 
   return app